Open index delegator whenever new SearchTabManager is created
Protege re-creates the SearchTabManager whenever any preferences are changed, regardless of which ones,
which is kind of odd. The index delegator is usually created when the ontology is changed, typically on
open. Adding a call to initIndex in the constuctor won't usually do anything as an ontology is not open
when it's created, but in this edge case of a preferences change it will re-create the index delegator.