Fix edge case in computing intersections of result sets
The common utility for computing intersections is incorrect. It assumes that if the first set is empty
that this is the first time called and it adds all the results from the second set to the first. It also assumes
the first is the running final results set from a collection of filters. This fix removes the need to call the utility
function.